Move the delete group methods into a separate extension like the other actions
